Refactor: MPAM-related code into val/common
 - Moved MPAM-related code into val/common
 - Relocated HMAT, PPTT, and SRAT implementations to val/common
 - Applied corresponding adjustments in pal to reflect structural changes
