+++ pal/uefi_acpi/SbsaPalLib.inf
@@ -1,5 +1,5 @@
 ## @file
-# Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+# Copyright (c) 2024-2025, Arm Limited or its affiliates. All rights reserved.
 #  SPDX-License-Identifier : Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -39,15 +39,14 @@
   common/src/pal_peripherals.c
   common/src/pal_exerciser.c
   common/src/pal_smmu.c
+  common/src/pal_mpam.c
+  common/src/pal_hmat.c
+  common/src/pal_pptt.c
+  common/src/pal_pcc.c
   sbsa/src/pal_sbsa_acpi.c
-  sbsa/src/pal_sbsa_hmat.c
   sbsa/src/pal_sbsa_ras.c
-  sbsa/src/pal_sbsa_misc.c
-  sbsa/src/pal_sbsa_mpam.c
-  sbsa/src/pal_sbsa_pptt.c
   sbsa/src/pal_sbsa_pmu.c
   sbsa/src/pal_sbsa_exerciser.c
-  sbsa/src/pal_sbsa_pcc.c
 
 [Packages]
   ArmPkg/ArmPkg.dec
