+++ pal/uefi_acpi/common/src/pal_pcc.c
@@ -1,6 +1,6 @@
 
 /** @file
- * Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2025, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,6 @@
 #include "Include/IndustryStandard/Acpi65.h"
 
 #include "common/include/pal_uefi.h"
-#include "sbsa/include/pal_sbsa_uefi.h"
 
 #define ADD_PTR(t, p, l) ((t *)((UINT8 *)p + l))
