+++ val/bsa/src/bsa_acs_gic.c
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2024-2025, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,30 +23,6 @@
 #include "bsa/include/bsa_pal_interface.h"
 
 extern GIC_INFO_TABLE  *g_gic_info_table;
-
-/**
-  @brief   This function routes interrupt to specific PE.
-           1. Caller       -  Test Suite
-           2. Prerequisite -  val_gic_create_info_table
-  @param   int_id Interrupt ID to be routed
-  @param   mpidr MPIDR_EL1 reg value of the PE to which the interrupt should be routed
-  @return  status
-**/
-uint32_t val_gic_route_interrupt_to_pe(uint32_t int_id, uint64_t mpidr)
-{
-  uint64_t cpuaffinity;
-
-  if (int_id > 31) {
-      cpuaffinity = mpidr & (PE_AFF0 | PE_AFF1 | PE_AFF2 | PE_AFF3);
-      val_mmio_write64(val_get_gicd_base() + GICD_IROUTER + (8 * int_id), cpuaffinity);
-  }
-  else{
-      val_print(ACS_PRINT_ERR, "\n    Only SPIs can be routed, ", 0);
-      val_print(ACS_PRINT_ERR, "interrupt with INTID = %d cannot be routed", int_id);
-  }
-
-  return 0;
-}
 
 /**
   @brief   This function will return '1' if an interrupt is either pending or active.
