+++ val/common/include/acs_pe.h
@@ -113,6 +113,9 @@ typedef enum {
   SCTLR_EL2,
   SCTLR_EL1,
   PMCR_EL0,
+  PMCCNTR_EL0,
+  PMCCFILTR_EL0,
+  PMCNTENSET_EL0,
   PMOVSSET_EL0,
   PMOVSCLR_EL0,
   PMINTENSET_EL1,
