+++ val/sbsa/include/sbsa_val_interface.h
@@ -32,16 +32,9 @@
 #endif
 
 
-typedef enum {
-  PE_FEAT_MPAM,
-  PE_FEAT_PMU,
-  PE_FEAT_RAS
-} PE_FEAT_NAME;
-
 /* PE related APIs */
 uint32_t val_pe_get_index_uid(uint32_t uid);
 uint32_t val_pe_get_uid(uint64_t mpidr);
-uint32_t val_pe_feat_check(PE_FEAT_NAME pe_feature);
 
 #if defined(TARGET_LINUX) || defined(TARGET_EMULATION)
 uint32_t val_get_device_path(const char *hid, char hid_path[][MAX_NAMED_COMP_LENGTH]);
@@ -60,56 +53,8 @@ uint32_t val_nist_generate_rng(uint32_t *rng_buffer);
 void     val_pmu_create_info_table(uint64_t *pmu_info_table);
 void     val_pmu_free_info_table(void);
 
-/*Cache related info APIs*/
-#define INVALID_CACHE_INFO 0xFFFFFFFFFFFFFFFF
-#define CACHE_TABLE_EMPTY 0xFFFFFFFF
-#define DEFAULT_CACHE_IDX 0xFFFFFFFF
 
-typedef enum {
-  CACHE_TYPE,
-  CACHE_SIZE,
-  CACHE_ID,
-  CACHE_NEXT_LEVEL_IDX,
-  CACHE_PRIVATE_FLAG
-} CACHE_INFO_e;
-
-void val_cache_create_info_table(uint64_t *cache_info_table);
-void val_cache_free_info_table(void);
-uint64_t val_cache_get_info(CACHE_INFO_e type, uint32_t cache_index);
-uint32_t val_cache_get_llc_index(void);
-uint32_t val_cache_get_pe_l1_cache_res(uint32_t res_index);
 uint64_t val_get_primary_mpidr(void);
-
-/* MPAM tests APIs */
-#define MPAM_INVALID_INFO 0xFFFFFFFF
-#define SRAT_INVALID_INFO 0xFFFFFFFF
-#define HMAT_INVALID_INFO 0xFFFFFFFF
-
-void val_mpam_create_info_table(uint64_t *mpam_info_table);
-void val_mpam_free_info_table(void);
-
-typedef enum {
-  MPAM_RSRC_TYPE_PE_CACHE,
-  MPAM_RSRC_TYPE_MEMORY,
-  MPAM_RSRC_TYPE_SMMU,
-  MPAM_RSRC_TYPE_MEM_SIDE_CACHE,
-  MPAM_RSRC_TYPE_ACPI_DEVICE,
-  MPAM_RSRC_TYPE_UNKNOWN = 0xFF  /* 0x05-0xFE Reserved for future use */
-} MPAM_RSRC_LOCATOR_TYPE;
-
-/* MPAM info request types*/
-typedef enum {
-  MPAM_MSC_RSRC_COUNT,
-  MPAM_MSC_RSRC_RIS,
-  MPAM_MSC_RSRC_TYPE,
-  MPAM_MSC_BASE_ADDR,
-  MPAM_MSC_ADDR_LEN,
-  MPAM_MSC_RSRC_DESC1,
-  MPAM_MSC_RSRC_DESC2,
-  MPAM_MSC_NRDY,
-  MPAM_MSC_ID,
-  MPAM_MSC_INTERFACE_TYPE
-} MPAM_INFO_e;
 
 /* RAS APIs */
 #define INVALID_RAS2_INFO 0xFFFFFFFFFFFFFFFFULL
@@ -130,24 +75,6 @@ uint64_t val_ras2_get_mem_info(RAS2_MEM_INFO_e type, uint32_t index);
 
 /* ETE */
 uint32_t val_sbsa_ete_execute_tests(uint32_t level, uint32_t num_pe);
-
-/* HMAT APIs */
-void val_hmat_create_info_table(uint64_t *hmat_info_table);
-void val_hmat_free_info_table(void);
-
-/* SRAT APIs */
-typedef enum {
-  SRAT_MEM_NUM_MEM_RANGE,
-  SRAT_MEM_BASE_ADDR,
-  SRAT_MEM_ADDR_LEN,
-  SRAT_GICC_PROX_DOMAIN,
-  SRAT_GICC_PROC_UID,
-  SRAT_GICC_REMOTE_PROX_DOMAIN
-} SRAT_INFO_e;
-
-void val_srat_create_info_table(uint64_t *srat_info_table);
-void val_srat_free_info_table(void);
-uint64_t val_srat_get_info(SRAT_INFO_e type, uint64_t prox_domain);
 
 #define PMU_INVALID_INFO 0xFFFFFFFFFFFFFFFF
 #define PMU_INVALID_INDEX 0xFFFFFFFF
@@ -176,11 +103,5 @@ uint32_t val_sbsa_pmu_execute_tests(uint32_t level, uint32_t num_pe);
 uint32_t val_sbsa_mpam_execute_tests(uint32_t level, uint32_t num_pe);
 uint32_t val_sbsa_ras_execute_tests(uint32_t level, uint32_t num_pe);
 uint32_t val_sbsa_nist_execute_tests(uint32_t level, uint32_t num_pe);
-
-/* PCC related APIs */
-void val_pcc_create_info_table(uint64_t *pcc_info_table);
-void *val_pcc_cmd_response(uint32_t subspace_id, uint32_t command, void *data, uint32_t data_size);
-uint32_t val_pcc_get_ss_info_idx(uint32_t subspace_id);
-void val_pcc_free_info_table(void);
 
 #endif
